chore(components): remove aiplatform.CustomContainerTrainingJobRunOp and aiplatform.CustomPythonPackageTrainingJobRunOp from GCPC major version 2
PiperOrigin-RevId: 526489601
