+++ components/google-cloud/google_cloud_pipeline_components/experimental/forecasting/__init__.py
@@ -14,10 +14,7 @@
 """Google Cloud Pipeline Experimental Forecasting Components."""
 
 import os
-from typing import Optional
 
-from google.cloud import aiplatform as aiplatform_sdk
-from google_cloud_pipeline_components.aiplatform import utils
 
 from kfp.components import load_component_from_file
